Move event handler to events.js
-Was is api, moved to correct folder.
